Microbot

Documentation for Microbot.

Microbot.add_three

Microbot.detrend

Microbot.fftclean

Microbot.fit_line

Microbot.add_three — Method

add_three(x)

Compute x+3. Used as an example test function to ensure tests are working correctly.

Microbot.detrend — Method

detrend(y::AbstractVector)

Remove a linear trend in y.

Microbot.fftclean — Method

fftclean(x::AbstractVector, y::AbstractVector)

Transform time data x, y into the frequency domain. Returns xf, yf.

Microbot.fit_line — Method

fit_line(x::AbstractVector, y::AbstractVector)

Fit a linear equation to data x, y. Returns m, b.

$y = mx + b$

Dependent on Optim.