API

I see this as a section that contains the function docstrings for everything in the package.

I'll have to think about how to break it down into modules so its sorted here and easily digestible.

MicroTracker.add_info_columns_from_filename — Method

add_info_columns_from_filename(particle_data::AbstractDataFrame, translation_dict::AbstractDict)

Extract experimental metadata from the filename column and create a new column for each.

Assumes the filename is separated by underscores and periods are denoted by p. The translation_dict details the filename format. For full explanation, see the MicroTracker docs (ref needed).

Example

julia> test_filename = "5_13p5_61p35"
"5_13p5_61p35"

MicroTracker.add_resolution_column! — Method

add_resolution_column(particle_data::AbstractDataFrame)

Look at a frame of the video in original_video which corresponds to the data and add a column.

MicroTracker.add_useful_columns — Method

add_useful_columns(linked_data::AbstractDataFrame, linking_settings::NamedTuple)

After linking, add some useful columns to the dataframe.

These include dx, dy, dp (velocity), and size measurements in microns. (more detailed docs needed if this is an interface for users)

MicroTracker.batch_particle_data_to_linked_data — Method

batch_particle_data_to_linked_data(translation_dict::Dict, linking_settings::NamedTuple; save_to_csv=true)

Process all .csv files in particle_data into linked trajectory data and concatenate the results.

Returns a DataFrame containing all linked data for the entire experimental array. This is also saved to linked_data using save_linked_data_with_metadata for further analysis.

The translation_dict` is a dictionary detailing the information contained in the filename. For full explanation, see the MicroTracker docs (ref needed).

julia> 1 + 1
2

MicroTracker.clip_trajectory_edges — Method

clip_trajectory_edges(linked_data::AbstractDataFrame)

Iterate through each trajectory and remove the tracking data where the particle is out of frame.

The particle is out of frame when the center is within the radius of the particle from the edge of the video.

MicroTracker.collapse_data — Method

collapse_data(linked_data::AbstractDataFrame)

Collapse each time-series microbot trajectory into a single row of summary data for each microbot.

For full description of each output variable, see these docs (ref needed).

MicroTracker.create_imagej_macro_here — Method

create_imagej_macro_here(;MPP::Float64, minimum_segmentation_diameter::Float64)

Create a macro script for batch segmenting videos in original_video with Fiji.

Required keyword arguments:

MPP = microns per pixel ::Float64

minimum_segmentation_diameter = only particles above this diameter in microns will be segmented ::Float64

MicroTracker.create_project_here — Method

create_project_here(;include_examples=false)

Create the folder structure for a MicroTracker project in the current working directory.

If include_examples is true, then example particle data and videos will be included.

MicroTracker.detrend — Method

detrend(y::AbstractVector)

Remove a linear trend in y.

MicroTracker.ellipse — Method

ellipse(major, minor, pos, angle)

Get points x, y for the ellipse with major and minor axes, pos vector, and angle in radians.

MicroTracker.estimate_omega — Method

estimate_omega(x::AbstractVector, y::AbstractVector)

Get the highest frequency peak of x, y data and return the corresponding xf divided by 2.

This works for rolling microbots when the y data exhibits two peaks every rotation. Depending on the data, this could be the Angle, Major_m, or V columns.

MicroTracker.extract_frame_from_label — Method

extract_frame_from_label(label::AbstractString)

When using ImageJ data, the frame column does not exist, so we need to pull it out of the label column.

This function assumes the frame number is at the end of the label, and that it is the only number in the label.

MicroTracker.fftclean — Method

fftclean(x::AbstractVector, y::AbstractVector)

Transform time data x, y into the frequency domain. Returns xf, yf.

MicroTracker.filter_trajectories — Method

filter_trajectories(collapsed_data::AbstractDataFrame, filter_settings::NamedTuple)

Docs needed.

MicroTracker.find_relevant_FPS — Method

find_relevant_FPS(particle_data::AbstractDataFrame, linking_settings::NamedTuple)

Used to allow for the FPS to be specified in the filename, or in the linking_settings.

If FPS is in the particledata, it will be used. If FPS is in the linkingsettings, it will be used. If FPS is in both, an error will be thrown.

MicroTracker.find_trajectory_bounds — Method

find_trajectory_bounds(df_1particle::AbstractDataFrame, video_resolution::Tuple{Int, Int})

Return a tuple of frame numbers, (low, high) where all trajectory points are in bounds.

This calculates the radius of the particle, then iterates forward and backward from the center of the trajectory until it finds a point that is out of bounds. This is the high and low bound of the trajectory.

MicroTracker.fit_line — Method

fit_line(x::AbstractVector, y::AbstractVector)

Fit a linear equation to data x, y. Returns m, b.

$y = mx + b$

Dependent on Optim.

MicroTracker.get_assets_path — Method

get_asset_path()

Always returns the path to the assets folder in the package root directory.

This works from the dev's, test suite's, or user's perspective.

MicroTracker.get_names_in_particle_data — Method

get_names_in_particle_data()

Return a list of all .csv files contained in the particle_data folder.

MicroTracker.getfileextension — Method

getfileextension(imagename::AbstractString)

Get the file extension from the image name.

Example

julia> MicroTracker.getfileextension("a0001.tif")
"tif"

MicroTracker.getframenumfromimagename — Method

getframenumfromimagename(imagename::AbstractString)

Get the relevant frame number from the name of an image. Does not necessarily match the real frame number in the data.

Example

julia> MicroTracker.getframenumfromimagename("5 24p9 1 2022-04-04-16-30-31-16-a-985.tif")
985

julia> MicroTracker.getframenumfromimagename("05_23_8_2 kept stack0001.tif")
1

MicroTracker.getimagenamedigitlength — Method

getimagenamedigitlength(tifname::AbstractString)

Get the length of the digits quantifying the frame in the image name. Used to catch leading zeros.

Example

julia> MicroTracker.getimagenamedigitlength("05_23_8_2 kept stack0001.tif")
4

MicroTracker.getprefixfromimagename — Method

getprefixfromimagename(imagename::AbstractString)

Return the prefix from the name of an image. This is everything besides the <framenumbers>.tif.

Example

julia> MicroTracker.getprefixfromimagename("5 24p9 1 2022-04-04-16-30-31-16-a-985.tif")
"5 24p9 1 2022-04-04-16-30-31-16-a-"

julia> MicroTracker.getprefixfromimagename("05_23_8_2 kept stack0001.tif")
"05_23_8_2 kept stack"

MicroTracker.getvideoresolution — Method

getvideoresolution(video_name::AbstractString)

Load in 1 frame of video_name and return the resolution of the video as a Tuple{Int, Int}

MicroTracker.inbounds — Method

inbounds(row::DataFrameRow, radius::AbstractFloat, video_resolution::Tuple{Int, Int})

Check if dataframe row is in bounds. This is a single row in a dataframe of a single particle's trajectory.

This looks at x and y coordinates and checks if they are within radius of the edge of the video.

MicroTracker.jldf_to_pydf — Method

jldf_to_pydf(jldf)

Convert a DataFrame (from DataFrames.jl) to a pandas.DataFrame. Used internally to interface with trackpy.

MicroTracker.link — Method

link(particle_data::AbstractDataFrame, linking_settings::NamedTuple)

Use trackpy to link particle data into trajectories across frames.

linking_settings is a NamedTuple with the following all caps fields:

SEARCH_RANGE_MICRONS::Float64: microns/s. Fastest a particle could be traveling.

MPP::Float64: Microns per pixel, scale of objective.

FPS::Float64: Frames per second. Can omit if the FPS has been specified in the filename.

STUBS_SECONDS::Float64: seconds. All trajectories that exist less than this will be removed.

MEMORY::Int64: Number of frames the particle can dissapear and still be remembered.

MicroTracker.load_linked_data — Method

load_linked_data(filename)

Read a linked data .csv file into a DataFrame. Assumes the .csv file is in the linked_data folder.

MicroTracker.load_particle_data — Method

load_particle_data(video_name::AbstractString)

Read a particle data .csv into a DataFrame. Assumes the .csv file is in the particle_data folder.

If the .csv file is from ImageJ, it will:

Extract the frame number from the label column and add it as a new column.

Rename the X and Y columns to x and y so it works with trackpy.

Remove any blank columns.

MicroTracker.loadframe — Method

loadframe(vidname::AbstractString, framenumber::Integer)

Return an image corresponding to the framenumber frame in the video vidname. Must be present in the original_video folder.

It looks at the way the tifs are automatically named and matches the pattern. Does not necessarily match the real frame number in the data, especially if there is no particle in the first frame.

MicroTracker.numerical_derivative — Method

numerical_derivative(x::AbstractVector)

Return the derivative of a Vector with spacing h = 1. Use on x and y columns of tracking data. Similar to Python's np.gradient function, but 61x faster here.

MicroTracker.particle_data_to_linked_data — Method

particle_data_to_linked_data(video_name::AbstractString, translation_dict::Dict, linking_settings::NamedTuple)

Process particle data into linked trajectory data while calculating instantaneous velocity and other salient data. Returns a DataFrame that can then be saved to linked_data using save_linked_data_with_metadata.

A particle data csv corresponding to video_name must be present in the particle_data folder. The translation_dict` is a dictionary detailing the information contained in the filename. For full explanation, see the MicroTracker docs (ref needed).

MicroTracker.plotannotatedframe — Method

plotannotatedframe(linked_data::AbstractDataFrame, filename::AbstractString, framenumber::Int; showimage=true, showellipse=true, plotkwargs...)

Display a single frame of a video with all microbot trajectories overlaid. Optionally, hide the fit ellipses or the image.

To plot a single microbot's trajectory, use plotannotatedframe_single. plotkwargs are passed to plot!() and are intended to change the appearance of the plot.

MicroTracker.plotannotatedframe_single — Method

plotannotatedframe_single(linked_data::AbstractDataFrame, particle_unique::AbstractString, framenumber::Int; showimage=true, showellipse=true, plotkwargs...)

Display a single frame of a video, with a single chosen microbot trajectory overlaid. Optionally, hide the fit ellipses or the image.

To plot all microbot trajectories, use plotannotatedframe. plotkwargs are passed to plot!() and are intended to change the appearance of the plot.

MicroTracker.pydf_to_jldf — Method

pydf_to_jldf(pydf)

Convert a pandas.DataFrame to a DataFrame (from DataFrames.jl). Used internally to interface with trackpy.

MicroTracker.save_linked_data_with_metadata — Method

save linked data with metadata

MicroTracker.total_displacement — Method

total_displacement(x::AbstractVector, y::AbstractVector)

Return the total displacement of the particle in pixels.

MicroTracker.trajectory_analyzer — Method

trajectory_analyzer(linked_data::AbstractDataFrame, collapsed_data::AbstractDataFrame, particle_unique::AbstractString; size_variable="Major_um", annotationkwargs...)

Display a comphrehensive dashboard of a single microbot, including instantaneous velocity, a chosen size_variable, and the FFT of the size_variable.

This version is static, i.e. the last frame is chosen for the video display. For an interactive version for use in a Pluto notebook, use trajectory_analyzer_scrubbable.

MicroTracker.trajectory_analyzer_scrubbable — Method

trajectory_analyzer_scrubbable(linked_data::AbstractDataFrame, collapsed_data::AbstractDataFrame, particle_unique::AbstractString, framenumber::Int; size_variable="Major_um", annotationkwargs...)

Display a comphrehensive dashboard of a single microbot, including instantaneous velocity, a chosen size_variable, and the FFT of the size_variable.

This version is scrubbable, i.e. a certain frame can be chosen for the video display and highlighted data point. This is highly recommended for use in a Pluto notebook. annotationkwargs are passed to plotannotatedframe_single().

To see a static last frame version of this plot, use trajectory_analyzer.

MicroTracker.translation_dict_to_string — Method

translation_dict_to_string(translation_dict::Dict)

Convert a dictionary into a string that can be used as a filename. Used internally to save linked_data.

Example

julia> test_translation_dict = Dict("f_Hz"=>(1, Int64), "B_mT"=>(2, Float64), "FPS"=>(3, Float64));

julia> MicroTracker.translation_dict_to_string(test_translation_dict)
"B_mT=(2, Float64), FPS=(3, Float64), f_Hz=(1, Int64)"