API

I see this as a section that contains the function docstrings for everything in the package.

I'll have to think about how to break it down into modules so its sorted here and easily digestible.

MicroTracker.collapse_time_data — Method

collapse_time_data(df)

Collapse each time-series microbot trajectory into a single row of data for each microbot.

For full description of each output variable, see these docs (ref needed).

MicroTracker.detrend — Method

detrend(y::AbstractVector)

Remove a linear trend in y.

MicroTracker.estimate_omega — Method

estimate_omega(x::AbstractVector, y::AbstractVector)

Get the highest frequency peak of x, y data and return the corresponding xf divided by 2.

This works for rolling microbots when the y data exhibits two peaks every rotation. Depending on the data, this could be the Angle, Major_m, or V columns.

MicroTracker.fftclean — Method

fftclean(x::AbstractVector, y::AbstractVector)

Transform time data x, y into the frequency domain. Returns xf, yf.

MicroTracker.fit_line — Method

fit_line(x::AbstractVector, y::AbstractVector)

Fit a linear equation to data x, y. Returns m, b.

$y = mx + b$

Dependent on Optim.

MicroTracker.get_particle_csvs — Method

get_particle_csvs()

Return a list of all .csv files contained in the particle_data folder.

MicroTracker.npsin — Method

npsin(a)

Placeholder test function that wraps np.sin from Python's numpy library.

MicroTracker.read_linked_csv — Method

read_linked_csv(filename)

Read a linkedcsv into a DataFrame. Assumes the .csv file is in the `linkedresults` folder.